# [NUI] Property setter is changed