### Description of Change ###
[NUI] Property setter is changed

### API Changes ###
N/A