@dongsug-song 2019-10-21
@huiyueun 
@adeelkazmi 
Hello, 
please review this. I tested this with "test/NUITestSample/NUITestSample/examples/PropertyCascadeCheckTest.cs"
on MuseM TV successfully.
thank you.