@adeelkazmi 2019-10-21
Sorry, I don't understand why we're doing this. Could you please explain why we're using a method in the setters now? It seems like it's private as well so deriving classes will not be able to override it.

It just seems like a lot of boiler plate.