@dongsug-song 2019-10-22
Hello, Mr. Adeel
I did some speed test with **_PropertyCascadeSpeedTest.cs_** on MuseM TV target.
please check the below.
thank you.

* AOT disabled and use locally built Tizen.NUI.dll for testing.
- propertySet1 step: set random values into properties.
- propertySet2 step: read the properties and change their values and set them again.
- propertySet3 step: change properties with cascade chaining by using '+='. In this case, current NUI is not working properly.


**Before this patch (current NUI), process time (small is better)**
propertySet1 1st : 204.2 ms 
propertySet1 2nd : 1.3 ms
propertySet1 3rd : 0.8 ms

propertySet2 1st : 34.6 ms
propertySet2 2nd : 2.0 ms
propertySet2 3rd : 1.1 ms

propertySet3 1st : 17.0 ms
propertySet3 2nd : 1.4 ms
propertySet3 3rd : 1.4 ms

**After this patch (locally built NUI), process time (small is better)**
propertySet1 1st : 203.4 ms
propertySet1 2nd : 1.6 ms
propertySet1 3rd : 0.6 ms

propertySet2 1st : 35.6 ms
propertySet2 2nd : 1.2 ms
propertySet2 3rd : 1.0 ms

propertySet3 1st : 22.5 ms
propertySet3 2nd : 4.0 ms
propertySet3 3rd : 4.2 ms



* propertySet1 and 2 's performances are almost same. the 3 is different(After-Patch one is bad). However, I think that Before-Patch is impossible to set cascade chaining property, so be NO meaningful.

* I think we need to apply this, because usually c# developers consider this kind of cascade chaining set as obvious and natural stuff.


**raw data (logs)**
F/NUITEST (22549): propertySet-speedtest.cs: setProperty1(107) > setProperty1() elapsed time=223.104ms
F/NUITEST (22549): propertySet-speedtest.cs: setProperty1(107) > setProperty1() elapsed time=1.0368ms
F/NUITEST (22549): propertySet-speedtest.cs: setProperty1(107) > setProperty1() elapsed time=0.96ms
F/NUITEST (22549): propertySet-speedtest.cs: setProperty2(215) > setProperty2() elapsed time=53.5808ms
F/NUITEST (22549): propertySet-speedtest.cs: setProperty2(215) > setProperty2() elapsed time=1.856ms
F/NUITEST (22549): propertySet-speedtest.cs: setProperty2(215) > setProperty2() elapsed time=1.3696ms
F/NUITEST (22549): propertySet-speedtest.cs: setProperty3(229) > property cascading set patch is NOT applied!
F/NUITEST (22549): propertySet-speedtest.cs: setProperty3(307) > setProperty3() elapsed time=31.872ms
F/NUITEST (22549): propertySet-speedtest.cs: setProperty3(229) > property cascading set patch is NOT applied!
F/NUITEST (22549): propertySet-speedtest.cs: setProperty3(307) > setProperty3() elapsed time=2.3808ms
F/NUITEST (22549): propertySet-speedtest.cs: setProperty3(229) > property cascading set patch is NOT applied!
F/NUITEST (22549): propertySet-speedtest.cs: setProperty3(307) > setProperty3() elapsed time=2.1376ms

F/NUITEST (30371): propertySet-speedtest.cs: setProperty1(107) > setProperty1() elapsed time=194.4448ms
F/NUITEST (30371): propertySet-speedtest.cs: setProperty1(107) > setProperty1() elapsed time=2.0736ms
F/NUITEST (30371): propertySet-speedtest.cs: setProperty1(107) > setProperty1() elapsed time=0.9088ms
F/NUITEST (30371): propertySet-speedtest.cs: setProperty2(215) > setProperty2() elapsed time=25.856ms
F/NUITEST (30371): propertySet-speedtest.cs: setProperty2(215) > setProperty2() elapsed time=3.0976ms
F/NUITEST (30371): propertySet-speedtest.cs: setProperty2(215) > setProperty2() elapsed time=1.1904ms
F/NUITEST (30371): propertySet-speedtest.cs: setProperty3(229) > property cascading set patch is NOT applied!
F/NUITEST (30371): propertySet-speedtest.cs: setProperty3(307) > setProperty3() elapsed time=9.9968ms
F/NUITEST (30371): propertySet-speedtest.cs: setProperty3(229) > property cascading set patch is NOT applied!
F/NUITEST (30371): propertySet-speedtest.cs: setProperty3(307) > setProperty3() elapsed time=0.9984ms
F/NUITEST (30371): propertySet-speedtest.cs: setProperty3(229) > property cascading set patch is NOT applied!
F/NUITEST (30371): propertySet-speedtest.cs: setProperty3(307) > setProperty3() elapsed time=1.0752ms


F/NUITEST (31915): propertySet-speedtest.cs: setProperty1(107) > setProperty1() elapsed time=195.3152ms
F/NUITEST (31915): propertySet-speedtest.cs: setProperty1(107) > setProperty1() elapsed time=0.8832ms
F/NUITEST (31915): propertySet-speedtest.cs: setProperty1(107) > setProperty1() elapsed time=0.6016ms
F/NUITEST (31915): propertySet-speedtest.cs: setProperty2(215) > setProperty2() elapsed time=24.3968ms
F/NUITEST (31915): propertySet-speedtest.cs: setProperty2(215) > setProperty2() elapsed time=1.216ms
F/NUITEST (31915): propertySet-speedtest.cs: setProperty2(215) > setProperty2() elapsed time=1.0112ms
F/NUITEST (31915): propertySet-speedtest.cs: setProperty3(229) > property cascading set patch is NOT applied!
F/NUITEST (31915): propertySet-speedtest.cs: setProperty3(307) > setProperty3() elapsed time=9.3568ms
F/NUITEST (31915): propertySet-speedtest.cs: setProperty3(229) > property cascading set patch is NOT applied!
F/NUITEST (31915): propertySet-speedtest.cs: setProperty3(307) > setProperty3() elapsed time=1.0368ms
F/NUITEST (31915): propertySet-speedtest.cs: setProperty3(229) > property cascading set patch is NOT applied!
F/NUITEST (31915): propertySet-speedtest.cs: setProperty3(307) > setProperty3() elapsed time=1.0368ms




F/NUITEST ( 5923): propertySet-speedtest.cs: setProperty1(107) > setProperty1() elapsed time=209.4592ms
F/NUITEST ( 5923): propertySet-speedtest.cs: setProperty1(107) > setProperty1() elapsed time=1.8944ms
F/NUITEST ( 5923): propertySet-speedtest.cs: setProperty1(107) > setProperty1() elapsed time=0.704ms
F/NUITEST ( 5923): propertySet-speedtest.cs: setProperty2(215) > setProperty2() elapsed time=36.6592ms
F/NUITEST ( 5923): propertySet-speedtest.cs: setProperty2(215) > setProperty2() elapsed time=1.2672ms
F/NUITEST ( 5923): propertySet-speedtest.cs: setProperty2(215) > setProperty2() elapsed time=0.9216ms
F/NUITEST ( 5923): propertySet-speedtest.cs: setProperty3(225) > property cascading set patch is applied!
F/NUITEST ( 5923): propertySet-speedtest.cs: setProperty3(307) > setProperty3() elapsed time=25.1136ms
F/NUITEST ( 5923): propertySet-speedtest.cs: setProperty3(225) > property cascading set patch is applied!
F/NUITEST ( 5923): propertySet-speedtest.cs: setProperty3(307) > setProperty3() elapsed time=3.84ms
F/NUITEST ( 5923): propertySet-speedtest.cs: setProperty3(225) > property cascading set patch is applied!
F/NUITEST ( 5923): propertySet-speedtest.cs: setProperty3(307) > setProperty3() elapsed time=4.1216ms




F/NUITEST ( 6118): propertySet-speedtest.cs: setProperty1(107) > setProperty1() elapsed time=198.848ms
F/NUITEST ( 6118): propertySet-speedtest.cs: setProperty1(107) > setProperty1() elapsed time=2.2528ms
F/NUITEST ( 6118): propertySet-speedtest.cs: setProperty1(107) > setProperty1() elapsed time=0.7168ms
F/NUITEST ( 6118): propertySet-speedtest.cs: setProperty2(215) > setProperty2() elapsed time=35.968ms
F/NUITEST ( 6118): propertySet-speedtest.cs: setProperty2(215) > setProperty2() elapsed time=1.1648ms
F/NUITEST ( 6118): propertySet-speedtest.cs: setProperty2(215) > setProperty2() elapsed time=1.1008ms
F/NUITEST ( 6118): propertySet-speedtest.cs: setProperty3(225) > property cascading set patch is applied!
F/NUITEST ( 6118): propertySet-speedtest.cs: setProperty3(307) > setProperty3() elapsed time=22.8992ms
F/NUITEST ( 6118): propertySet-speedtest.cs: setProperty3(225) > property cascading set patch is applied!
F/NUITEST ( 6118): propertySet-speedtest.cs: setProperty3(307) > setProperty3() elapsed time=4.608ms
F/NUITEST ( 6118): propertySet-speedtest.cs: setProperty3(225) > property cascading set patch is applied!
F/NUITEST ( 6118): propertySet-speedtest.cs: setProperty3(307) > setProperty3() elapsed time=3.7248ms



F/NUITEST ( 6402): propertySet-speedtest.cs: setProperty1(107) > setProperty1() elapsed time=202.0224ms
F/NUITEST ( 6402): propertySet-speedtest.cs: setProperty1(107) > setProperty1() elapsed time=0.8576ms
F/NUITEST ( 6402): propertySet-speedtest.cs: setProperty1(107) > setProperty1() elapsed time=0.6144ms
F/NUITEST ( 6402): propertySet-speedtest.cs: setProperty2(215) > setProperty2() elapsed time=34.432ms
F/NUITEST ( 6402): propertySet-speedtest.cs: setProperty2(215) > setProperty2() elapsed time=1.3568ms
F/NUITEST ( 6402): propertySet-speedtest.cs: setProperty2(215) > setProperty2() elapsed time=1.2672ms
F/NUITEST ( 6402): propertySet-speedtest.cs: setProperty3(225) > property cascading set patch is applied!
F/NUITEST ( 6402): propertySet-speedtest.cs: setProperty3(307) > setProperty3() elapsed time=19.7504ms
F/NUITEST ( 6402): propertySet-speedtest.cs: setProperty3(225) > property cascading set patch is applied!
F/NUITEST ( 6402): propertySet-speedtest.cs: setProperty3(307) > setProperty3() elapsed time=3.6224ms
F/NUITEST ( 6402): propertySet-speedtest.cs: setProperty3(225) > property cascading set patch is applied!
F/NUITEST ( 6402): propertySet-speedtest.cs: setProperty3(307) > setProperty3() elapsed time=4.9024ms